+++ scripts/settings.js
@@ -18,6 +18,7 @@ import {actions} from './extensions/actions.js';
 import {item} from './applications/item.js';
 import {activities} from './extensions/activities.js';
 import {combat} from './extensions/combat.js';
+import {ui} from './extensions/ui.js';
 function addSetting(options) {
     let setting = {
         scope: options.scope ?? 'world',
@@ -883,7 +884,8 @@ export function registerSettings() {
         type: Boolean,
         default: false,
         category: 'development',
-        onChange: () => activities.cssTweak()
+        onChange: (value) => { activities.cssTweak(value);
+        }
     });
     addSetting({
         key: 'groupSummonsWithOwner',
@@ -912,6 +914,27 @@ export function registerSettings() {
                 Hooks.off('preUpdateCombat', combat.legendaryActionsPrompt);
             }
         }
+    });
+    addSetting({
+        key: 'customUIButtonScale',
+        type: Number,
+        default: 1,
+        category: 'interface',
+        onChange: (value) => ui.setBodyProperty('--custom-ui-button-scale', value)
+    });
+    addSetting({
+        key: 'customUINavigationScale',
+        type: Number,
+        default: 1,
+        category: 'interface',
+        onChange: (value) => ui.setBodyProperty('--custom-ui-navigation-scale', value)
+    });
+    addSetting({
+        key: 'customSidebar',
+        type: Boolean,
+        default: false,
+        category: 'interface',
+        onChange: (value) => ui.customSidebar(value)         
     });
 }
 export function registerMenus() {